Protobuf support (#92)
* Also add the dependencies in /examples/WORKSPACE

This allow to runs the tests in examples as a remote repository
and as a local repository.

* Add a dummy Cargo craze to use cargo raze on it.

With this craze, and cargo raze, we can simply use raze generated repositories
to add the necessary dependencies.

* Add generated files for the raze package of protobuf

* Add rust_(proto|grpc)_library

Those library add protobuf / gRPC support to Rust language in Bazel
using the native proto_library. It generates one crate per proto_library
which is efficient for incrementality but can lead to naming collisions.

We expose a bit of rust_library internal to re-use it in rust_proto_library
and create a new provider "rust_libs" that will declare a group of rust libraries.

* Add a gRPC client/server example

And a test with it.

* Add documentation for the rust_(proto|grpc)_library

* Use the new runfiles library in the test rather than a custom algorithm

* Add the missing dependencies to examples/WORKSPACE

This new dependency is quite unfortunate, it make using rules_rust
more complex since it needs to import a bzl file from skylib to be
able to use rules_rust, that adds 10 lines in the WORKSPACE without
much improvement. I guess this cannot be better until recursive
workspace are a thing.

* Generates cargo raze files compatible with macOS and Linux

Also to keep the same lockfile, check in the Cargo.lock and have
a lever to not re-run cargo generate-lockfile.

* Compile all proto stubs in one library

This a workaround a bug in Bazel that prevent the toolchain from being
loaded with the target transition (see
https://github.com/bazelbuild/bazel/pull/5650).

This is not ideal because it recompiles several time the same
proto if you depends on it from several dependency path
and can lead to seeing two compilation of the same proto
as different type. This is also terrible for incremental build
and contrary to the (protobuf blog
entry)[https://blog.bazel.build/2017/02/27/protocol-buffers.html].

* Force set the lazy_static implementation configuration.

The lazy_static crates now set a configuration from the build.rs, ideally we would be
able to handle that but that is not the case with the current cargo raze method.

* Final touch: documenting and simplifying toolchain creation

  - Documents how to create your custom toolchain when you need to import
    your own crates (or to use a different protobuf/grpc stubs compiler).
  - Add extra_aliased_targets to simply the cargo-raze generation
  - Use skylark http_archive to import protobuf.
  - Regenerate crates with google/cargo-raze#77 so that we don't need to
    have custom rewriting policy. Hence, deleted the wrapper around
    cargo raze, an update to the build file should simply be
    `cargo genrate-lockfiles && cargo raze` now.

* Fix nits suggested by @mfarrugi

* Remove usage of CrateInfos

This was a leftover from compiling the crates in the proto aspect.

* Use python for the optional_output_wrapper

Also move the optional_output_wrapper from the toolchain to the library
itself.
For some reason the runfiles where not shipped to the actions with the
old way which wasn't important for Bash that did not had any runfiles
but it is important for python.

* fix: README format

* Second pass of nits from @mfarrugi

* fix: add back the list of grpc stubs to create if not created