Fix rust_doc_test with transitive dependencies and add tests for it (#155)
* Test for failure to import crate in rust_doc_test.

This test fails with:
error[E0463]: can't find crate for
 --> test/chained_direct_deps/mod3.rs:4:1
  |
4 | extern crate mod1;
  | ^^^^^^^^^^^^^^^^^^ can't find crate

A git bisect indicate that #122 is the culprit change.

* fix: doc_test

Before #122 the test was wrongly passing.

* Fix rustdoc_test by using the dependency's providers

Instead of trying to recreate the DepInfo provider, use directly
the dep's DepInfo. It allows to correctly add all the direct
deps to the command line.

* Exclude rust_doc_test from RBE

All other rust_doc_test are already excluded.