feat: Support downloading nightly rustc-dev components (#349)
Working with the Rust compiler crates requires the rustc-dev component. This is distributed separately from the standard library. This PR adds a new option to download the rustc-dev component which defaults to False. The component is only distributed with nightly Rust, so the version option must be set to "nightly".