Pass extraneous environment variables to build scripts (#362)
cargo-raze has `buildrs_additional_environment_variables` which allows
extraneous environment variable to be passed to the build script. In addition
this should enable for everybody to add the environment variables that are
not added by the `cargo_build_script` rule itself.

That's the rules_rust side of the fix for google/cargo-raze#179.