use -msvc variant on Windows (#425)
Both the standard Rust installer, and rust_repositories() in this repo
default to the -msvc platform. Since it is the default on Windows,
this change comments out the -gnu variant instead.

This change (along with some fixes in cargo raze) allows crates with
msvc-specific build deps to compile. I do not know enough about Bazel
to say if it will have any unintended side effects, but it seems to
work for me.