Remove uses of `BAZEL_VERSION` for detection of pre-1.0 Bazel. (#508)
This PR removes:

* The special cases for pre-1.0 Bazel versions in `rustc.bzl`
* `BAZEL_VERSION` and `@ io_bazel_rules_rust_bazel_version`, which no longer have any references.
* `rust_workspace()` in `//:workspace.bzl`, which was obsolete due to no longer needing to define the Bazel version repository.

The remaining behavior of `rust_workspace()` was to define toolchains for Starlark unit testing, which are used by `rules_rust` itself but not necessary for downstream consumers. I moved that part to `//:WORKSPACE`.

For compatibility, I left the `rust_workspace()` symbol defined but empty (except for a warning message). It ought to be deleted in a future PR once users have removed it from their `WORKSPACE` files.

Fixes https://github.com/bazelbuild/rules_rust/issues/462