ccinfo: when providing ccinfo, optionally include libstd and alloc (#624)
* ccinfo: when providing ccinfo, optionally include libstd and alloc

The new attribute on RustToolchain is the label of a target that
provides __rust_realloc et al, which allows ld(1) to use the .rlib
files directly without needing to involve rustc in the linking
step. This means Rust and C++ can be mixed in a cc_binary freely
without needing any staticlib-type crates, which avoids problems if
you have a cc_binary -> rust_library -> cc_library -> rust_library
situation.

* Add verification that all files were properly partitioned

* Disable print

* stdlib_ordering: new tests to ensure check stdlib rlib ordering

* native_deps: give up on testing cc_binary using rust_library

This only works on clang today, and can only work on clang for the
near-term. It's not clear how to write a clang-only test in this case,
so we just give up for now.

"I promise it works!" etc

Co-authored-by: Augie Fackler <augie@google.com>
Co-authored-by: Marcel Hlopko <hlopko@google.com>