Fix rust-project.json generation in subpackages (#724)
This PR fixes a bug in `gen_rust_project` where we assumed that `rust-project.json` is generated directly under `bazel-bin`. For `rust_analyzer` targets deeper in the workspace tree that is not correct and this PR correctly descents into the right package in `bazel-bin` to find the `rust-project.json`.

I'm amazed how well I nerdsniped myself into writing a label parsing library in Rust. How hard can it be, though I. Foolish me.