Add tools attr with exec transition for cargo_build_script (#885)
If tools are passed in `data`, they must be built in the `exec` configuration in order to run on the host machine when the target != host.

I _think_ this should be a fairly safe change, unless someone is doing something with the tools passed to the build script that would require them to run on the target configuration, but short of copying these tools into the out_dir I don't know what that would be. And -- that would be better handled by depending directly on these tools in the data attr of the library or binary anyway.