Obtain the "CARGO_CRATE_NAME" value from the CrateInfo (#999)
At this point we have a `CrateInfo` for the target which contains the name of the crate, regardless whether it's been derived from the `ctx.attr.crate_name` or explicitly assigned. The code without this PR works well for the `rules_rust`, because the `CrateInfo.crate_name` is obtained by calling the `get_crate_name_from_attr` function. However, this is not sufficient for aspects, as they may be attached to a target that doesn't specify crate name, or one may want to assign a modified name to crates created by aspects.