Remove uses of `DepInfo.transitive_libs` (#1054)
This is a step further towards https://github.com/bazelbuild/rules_rust/issues/1051.

With this change, `rules_rust` doesn't need `DepInfo.transitive_libs` anymore. The only place where this field was used was to pass the files it contains as inputs to rust compile actions. We use `DepInfo.transitive_crate_outputs` for that now. For non-rlib compile actions, which need non-crate inputs as well, we add those inputs to the `additional_transitive_inputs` variable.