Enable xz archives (#1284)
* Expand the universe of known shas to include the .xz archive shas.

I have explicitly added keys for the files with the extension explicitly
on the file key. The point of this is to (in a future change) allow
tar.xz archives to be used instead of tar.gz archives.

This change keeps compatiblity with the old keys that assume the tar.gz
archives. I don't know if this is something that we should drop in the future
or just not worry about.

* Enable xz archive support.

XZ archives are about half the size of the gz equivalents. This
improves the cold repo bootstrap case a fair amount.

Co-authored-by: Daniel Wagner-Hall <dwagnerhall@apple.com>