Don't emit `CrateInfo` from `rust_static_library` and `rust_shared_library` (#1298)
These rules emit a `staticlib` and `cdylib` which are not meant for consumption by other rust targets.

However, one should still be able to write a `rust_test` for `rust_shared_library` and `rust_static_library` targets, which needs a `CrateInfo`. For this purpose we provide a `CrateInfo` wrapped in a `TestCrateInfo` provider, that `rust_test` understands.