Add support for aarch64-apple-ios-sim (#1304)
This adds / improves support for for the arm64 iOS simulator that's used
on M1 macs. Previously there was ambiguity in the constraints used for
this platform because bazel does not pick the best match, so this adds
the device constraint in this case to disambiguate. Ideally we could
rely on not having the `sim` suffix to be enough to add the device
constraint, but older triples like `x86_64-apple-ios` break this
pattern. I broke this logic out into a new function since it relies on
more info from the triple than just the ABI now.

https://github.com/bazelbuild/bazel/issues/11454

Co-authored-by: UebelAndre <github@uebelandre.com>