Build `rust_test` targets with `crate` using the same crate name as the underlying library target. (#1332)
This PR changes `rust_test` to use the same crate name as the underlying target, in cases where the `crate` attribute is used. This is consistent with `cargo`, which builds the crate using the name of the library when running `cargo test`.

The behavior of `rust_test` targets that don't use the `crate` attribute is unchanged.