Preserve directory structure of source files when some are generated (#1526)
When some input files are generated, the source files get symlinked to `bazel-out` so the generated and source files are in the same directory. This was implemented in #1340. However, the directory structure would get flattened since only the basename was used for the symlink location. Now the package-relative path is used.

Fixes #1510.