Yarn.lock changes should not cause a change in minor version.
Previously, I assumed that yarn.lock was used by npm on the server-side somehow for integrity checking. This is not the case, see: https://docs.npmjs.com/cli/v8/configuring-npm/package-lock-json

As a result, the lockfile minimization pass was completely superflous, and there was no need to publish a new minor version every time the lockfile changed.

The lockfile I believe is still needed to check for tampering during the publish step, but I am unsure of this.