Move crate_root_src to utils (#1570)
The honest motivation for this little refactoring is https://github.com/google/crubit. In build rules that don't yet work in the open-source yet we need to patch `rust.bzl` with a custom logic encapsulated in a function (the function collects extra dependencies from aspects). This function is called from `rust.bzl`, and it needs to call `crate_root_src`. To prevent a `load` cycle I'm moving the function to `utils.bzl`. I understand that it would be much better to justify this PR with a working build rules in Crubit, but that's unfortunately not easily achievable yet.