Fix build scripts targeting the wrong architecture (#1564)
* Fix build scripts targeting the wrong architecture

PR #1081 started passing CFLAGS to build scripts, which cargo does
too. But unfortunately Bazel's cpp toolchain adds an explicit target to
the CFLAGS it provides, and that target uses the host platform instead
of the target platform. When building a crate like blake3 or zstd on a
Mac with iOS as the target, this ends up in the crate being built for
macOS instead of iOS.

* Add an end-to-end test to make sure building zstd for iOS works