Incrementing version numbers for the whole monorepo.
Previously, deploy/program.ts generated a so-called 'syntheticVersion'
based off the Go versionless scheme.

Now a version number is generated from the files in VERSION.

This will break the deploy process once the first version is published,
because version bumping is not yet implemented for the whole monorepo.

In the future, it will be useful to add this version to package.json
and go.mod. I'm not sure of the correct mechanism there.