Fix tests accidentally ephemerally introduced in c466cd3a1697b51b7b7436be2095b26372abf460 and cfb413f7a0eafd6b6cbae2351eba227c2d07c592.
Prevent renovate adding all JSON files as part of automatic fixing.

Based on this comment, renovate was adding some very strange files https://github.com/Zemnmez/monorepo/pull/3488#issuecomment-1666254625 -- and as a result, the diff did NOT include an update to pnpm.lock, resulting in the tests passing when, since the lockfile was not updated, the changes did not happen correctly.