Deploy main to staging before trying to deploy the candidate branch.
This commit addresses a subtle, and fairly complex problem.

Pulumi does not know when two resources are mutually exclusive (e.g. DNS records in the same zone).

As a result, to correctly handle these cases, it must be made aware that these resources are the same via aliases.

However, a *failure* does not indicate a roll-back to a known good state -- meaning that under these conditions:

1. A mutually exclusive resource is deployed to the staging environment
2. That content is not destroyed
3. Another candidate branch attempts to deploy this mutually exclusive resource again

-- under these conditions, one candidate can fail due to a mistake in a previous test.

In order to thoroughly address this, we:

1. Destroy the Pulumi state after each presubmit to staging.
2. Before the next test, we deploy the origin/main state to the staging environment.

This ensures that the transition in infrastructure being tested is *always* the transition from main branch state to the PR / candidate branch state.

This ensures