Sketch of adding menu
This commit adds a menu to zemn.me. This, as it turns out is a bit more complicated than it seems at first.

## Column layout issues

It makes sense as part of this to move the 'About' section to its own page -- but the multicol layouts I'm trying to use
don't really work all that well, primarily becuase you can't really do 'pagination'.

By pagination, I mean the kind of breaks you'd imagine seeing on a multi-column document as it is printed on several sheets
of paper, like a whitepaper.

Because column layout allows no limit on the height of a column, the content inside a column can quickly grow so large that each column
is more than one page long. Then the user annoyingly has to scroll up and down to read all the text!!!

## Navbar issues

The navbar should probably have the site logo in, right? But the current hero area on the index page already has one. So if I put a header with a logo
in, we get two logos on the screen which looks super weird.

The alternative, which is some kind of custom navbar for the index page also seems weird. And removing the logo leaves a weird void.