Unprotect the callbox phone number.
I think I found the issue... I have been struggling with the fact that I can't seem to get the `outs` from my phone_number dynamic provider.

Trawling the docs made me realise that for each out to stay, it has to be declared in the *input args* param of the super constructor: https://www.pulumi.com/docs/iac/concepts/resources/dynamic-providers/#dynamic-resource-outputs

Transcluded here:

> export class MyResource extends pulumi.dynamic.Resource {
>     public readonly myStringOutput!: pulumi.Output<string>;
>     public readonly myNumberOutput!: pulumi.Output<number>;
>
>     constructor(name: string, props: MyResourceInputs, opts?: pulumi.CustomResourceOptions) {
>         super(new MyResourceProvider(), name, { myStringOutput: undefined, myNumberOutput: undefined, ...props }, opts);
>     }
> }

I did update this, but the outputs did not fix or update. I realise now this is because I *need to recreate the dynamic provider instance*. In order to do that I need to unprotect it.